requirements.txt
djangorestframework
psycopg2
djoser
pillow
channels
redis
channels_redis
